WIP: basic hex chapes, but jagged edges. Need to rework
